pyproject.toml
[build-system]
requires = [
  "setuptools>=68",
  "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "sciexplorer_v2"
version = "0.1.0"
description = "SciExplorer builds on LLM APIs to create artificial scientist agents that can explore unknown systems by controlling experiments and running analysis code. "
readme = "README.md"
requires-python = ">=3.14"
authors = [
  {name = "Maximilian Nägele"},
  {name = "Florian Marquardt"},
]

dependencies = [
  "arxiv==2.3.1",
  "feedparser==6.0.12",
  "ipython==9.7.0",
  "jax==0.8.1",
  "langchain==1.1.2",
  "langchain_classic==1.0.0",
  "langchain_community==0.4.1",
  "langchain_core==1.2.8",
  "langchain_google_genai==4.2.0",
  "langchain_openai==1.1.0",
  "langchain_text_splitters==1.0.0",
  "langgraph==1.0.4",
  "matplotlib==3.10.7",
  "numpy==2.3.5",
  "Pillow==12.0.0",
  "PyPDF2==3.0.1",
  "Requests==2.32.5",
  "scipy==1.16.3",
  "typing_extensions==4.15.0",
  "scikit-learn==1.7.2",
  "cloudpickle==3.1.2",
  "faiss-cpu==1.13.1"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["sciexplorer*"]
